update db structure

+++ itsybitsy/src/main/java/com/example/itsybitsy/Repository/EventsRepository.java
@@ -3,5 +3,5 @@ package com.example.itsybitsy.Repository;
 import com.example.itsybitsy.DbModels.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventsRepository extends JpaRepository<Event, Integer> {
+public interface EventsRepository extends JpaRepository<Event, Long> {
 }
